+++ cmd/starter/main_linux.go
@@ -66,12 +66,13 @@ func startup() {
 		sylog.Verbosef("Execute master process\n")
 
 		pid := sconfig.GetContainerPid()
+		imageFd := sconfig.GetImageFd()
 
 		if err := sconfig.Release(); err != nil {
 			sylog.Fatalf("%s", err)
 		}
 
-		starter.Master(int(C.rpc_socket[0]), int(C.master_socket[0]), pid, e)
+		starter.Master(int(C.rpc_socket[0]), int(C.master_socket[0]), int(C.cleanup_socket[0]), pid, imageFd, e)
 	case C.RPC_SERVER:
 		sylog.Verbosef("Serve RPC requests\n")
 
@@ -80,7 +81,14 @@ func startup() {
 		}
 
 		starter.RPCServer(int(C.rpc_socket[1]), e)
+	case C.CLEANUP_HOST:
+		sylog.Verbosef("Execute Cleanup Host Process")
+		if err := sconfig.Release(); err != nil {
+			sylog.Fatalf("%s", err)
+		}
+		starter.CleanupHost(int(C.cleanup_socket[1]), e)
 	}
+
 	sylog.Fatalf("You should not be there\n")
 }